+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - added Show task info button for quick access to task information
 - image messages are sent also when Include Context Files is disabled
 - properly initializing internal task in project
-- added Updated Files section with information about the changes
+- added Updated Files section with information about the changes and ability to request changes on specific line
 - BMAD mode consecutive prompts are now properly sending when working via browser
 - sync task baseDir with project baseDir to handle migrated tasks